requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
python-dotenv
sentence-transformers
jinja2
